# backend/src/main/java/io/reliza/ws/tea/ArtifactApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.21.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package io.reliza.ws.tea;

import io.reliza.model.tea.TeaArtifact;
import io.reliza.model.tea.TeaErrorResponse;
import java.util.UUID;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-03-29T10:44:15.267909500-04:00[America/Toronto]", comments = "Generator version: 7.21.0")
@Validated
@Tag(name = "TEA Artifact", description = "the TEA Artifact API")
public interface ArtifactApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    String PATH_GET_ARTIFACT_BY_VERSION = "/artifact/{uuid}/{artifactVersion}";
    /**
     * GET /artifact/{uuid}/{artifactVersion}
     * Get metadata for a specific revision of a specific TEA Artifact
     *
     * @param uuid UUID of TEA Artifact in the TEA server (required)
     * @param artifactVersion Version of TEA Artifact (required)
     * @return Requested TEA Artifact metadata found and returned (status code 200)
     *         or Request was Invalid (status code 400)
     *         or Object requested by identifier not found (status code 404)
     */
    @Operation(
        operationId = "getArtifactByVersion",
        description = "Get metadata for a specific revision of a specific TEA Artifact",
        tags = { "TEA Artifact" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Requested TEA Artifact metadata found and returned", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = TeaArtifact.class))
            }),
            @ApiResponse(responseCode = "400", description = "Request was Invalid"),
            @ApiResponse(responseCode = "404", description = "Object requested by identifier not found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = TeaErrorResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "basicAuth"),
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = ArtifactApi.PATH_GET_ARTIFACT_BY_VERSION,
        produces = { "application/json" }
    )
    default ResponseEntity<TeaArtifact> getArtifactByVersion(
        @Parameter(name = "uuid", description = "UUID of TEA Artifact in the TEA server", required = true, in = ParameterIn.PATH) @PathVariable("uuid") UUID uuid,
        @Parameter(name = "artifactVersion", description = "Version of TEA Artifact", required = true, in = ParameterIn.PATH) @PathVariable("artifactVersion") Integer artifactVersion
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"createdDate\" : \"\", \"formats\" : [ { \"checksums\" : [ { \"algValue\" : \"algValue\", \"algType\" : \"\" }, { \"algValue\" : \"algValue\", \"algType\" : \"\" } ], \"signatureUrl\" : \"http://example.com/aeiou\", \"description\" : \"description\", \"mediaType\" : \"mediaType\", \"url\" : \"http://example.com/aeiou\" }, { \"checksums\" : [ { \"algValue\" : \"algValue\", \"algType\" : \"\" }, { \"algValue\" : \"algValue\", \"algType\" : \"\" } ], \"signatureUrl\" : \"http://example.com/aeiou\", \"description\" : \"description\", \"mediaType\" : \"mediaType\", \"url\" : \"http://example.com/aeiou\" } ], \"name\" : \"name\", \"type\" : \"\", \"uuid\" : \"\", \"version\" : 6, \"distributionIds\" : [ null, null ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : \"OBJECT_UNKNOWN\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_GET_LATEST_ARTIFACT = "/artifact/{uuid}/latest";
    /**
     * GET /artifact/{uuid}/latest
     * Get metadata for latest revision of a specific TEA Artifact
     *
     * @param uuid UUID of TEA Artifact in the TEA server (required)
     * @return Requested TEA Artifact metadata found and returned (status code 200)
     *         or Request was Invalid (status code 400)
     *         or Object requested by identifier not found (status code 404)
     */
    @Operation(
        operationId = "getLatestArtifact",
        description = "Get metadata for latest revision of a specific TEA Artifact",
        tags = { "TEA Artifact" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Requested TEA Artifact metadata found and returned", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = TeaArtifact.class))
            }),
            @ApiResponse(responseCode = "400", description = "Request was Invalid"),
            @ApiResponse(responseCode = "404", description = "Object requested by identifier not found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = TeaErrorResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "basicAuth"),
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = ArtifactApi.PATH_GET_LATEST_ARTIFACT,
        produces = { "application/json" }
    )
    default ResponseEntity<TeaArtifact> getLatestArtifact(
        @Parameter(name = "uuid", description = "UUID of TEA Artifact in the TEA server", required = true, in = ParameterIn.PATH) @PathVariable("uuid") UUID uuid
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"createdDate\" : \"\", \"formats\" : [ { \"checksums\" : [ { \"algValue\" : \"algValue\", \"algType\" : \"\" }, { \"algValue\" : \"algValue\", \"algType\" : \"\" } ], \"signatureUrl\" : \"http://example.com/aeiou\", \"description\" : \"description\", \"mediaType\" : \"mediaType\", \"url\" : \"http://example.com/aeiou\" }, { \"checksums\" : [ { \"algValue\" : \"algValue\", \"algType\" : \"\" }, { \"algValue\" : \"algValue\", \"algType\" : \"\" } ], \"signatureUrl\" : \"http://example.com/aeiou\", \"description\" : \"description\", \"mediaType\" : \"mediaType\", \"url\" : \"http://example.com/aeiou\" } ], \"name\" : \"name\", \"type\" : \"\", \"uuid\" : \"\", \"version\" : 6, \"distributionIds\" : [ null, null ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : \"OBJECT_UNKNOWN\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}