# tea-spec/openapi.yaml
...
openapi: 3.1.1
...
info:
  title: Transparency Exchange API
  summary: The OWASP Transparency Exchange API specification for consumers and publishers
  description: TBC
  contact:
    name: TEA Working Group
    email: tbc@somewhere.tld
    url: https://github.com/CycloneDX/transparency-exchange-api
  license:
    name: Apache 2.0
    url: https://github.com/CycloneDX/transparency-exchange-api/blob/main/LICENSE
  version: 0.4.0
servers:
  - url: http://localhost/tea/v1
    description: Local development
paths:
...
  /component/{uuid}:
    get:
      description: Get a TEA Component
      operationId: getTeaComponentById
      parameters:
        - name: uuid
          in: path
          required: true
          description: UUID of TEA Component in the TEA server
          schema:
            "$ref": "#/components/schemas/uuid"
      responses:
        '200':
          description: Requested TEA Component found and returned
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/component"
        '400':
          $ref: "#/components/responses/400-invalid-request"
        '404':
          $ref: "#/components/responses/404-object-by-id-not-found"
      tags:
        - TEA Component
  /component/{uuid}/releases:
    get:
      description: Get releases of the component
      operationId: getReleasesByComponentId
      parameters:
        - name: uuid
          in: path
          required: true
          description: UUID of TEA Component in the TEA server
          schema:
            "$ref": "#/components/schemas/uuid"
      responses:
        '200':
          description: Requested Releases of TEA Component found and returned
          content:
            application/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/release"
        '400':
          $ref: "#/components/responses/400-invalid-request"
        '404':
          $ref: "#/components/responses/404-object-by-id-not-found"
      tags:
        - TEA Component
...
  /artifact/{uuid}/latest:
    get:
      description: Get metadata for latest revision of a specific TEA Artifact
      operationId: getLatestArtifact
      parameters:
        - name: uuid
          in: path
          required: true
          description: UUID of TEA Artifact in the TEA server
          schema:
            "$ref": "#/components/schemas/uuid"
      responses:
        '200':
          description: Requested TEA Artifact metadata found and returned
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/artifact"
        '400':
          $ref: "#/components/responses/400-invalid-request"
        '404':
          $ref: "#/components/responses/404-object-by-id-not-found"
      tags:
        - TEA Artifact
  /artifact/{uuid}/{artifactVersion}:
    get:
      description: Get metadata for a specific revision of a specific TEA Artifact
      operationId: getArtifactByVersion
      parameters:
        - name: uuid
          in: path
          required: true
          description: UUID of TEA Artifact in the TEA server
          schema:
            "$ref": "#/components/schemas/uuid"
        - name: artifactVersion
          in: path
          required: true
          description: Version of TEA Artifact
          schema:
            type: integer
      responses:
        '200':
          description: Requested TEA Artifact metadata found and returned
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/artifact"
        '400':
          $ref: "#/components/responses/400-invalid-request"
        '404':
          $ref: "#/components/responses/404-object-by-id-not-found"
      tags:
        - TEA Artifact
...
components:
  schemas:
...
    date-time:
      type: string
      description: Timestamp
      format: date-time
      pattern: "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}Z$"
      example: '2024-03-20T15:30:00Z'
    identifier:
      type: object
      description: An identifier with a specified type
      properties:
        idType:
          description: Type of identifier, e.g. `TEI`, `PURL`, `CPE`
          "$ref": "#/components/schemas/identifier-type"
        idValue:
          description: Identifier value
          type: string
    identifier-type:
      type: string
      description: Enumeration of identifiers types
      enum:
        - CPE
        - TEI
        - PURL
        - COMPLIANCE_DOCUMENT
...
    uuid:
      type: string
      description: A UUID
      format: uuid
      pattern: "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"

    #
    # TEA Product
    #
...
    component:
      type: object
      description: A TEA component
      properties:
        uuid:
          description: A unique identifier for the TEA component
          "$ref": "#/components/schemas/uuid"
        name:
          type: string
          description: Component name
        identifiers:
          type: array
          description: List of identifiers for the component
          items:
            "$ref": "#/components/schemas/identifier"
      required:
        - uuid
        - name
        - identifiers
      examples:
        - uuid: 3910e0fd-aff4-48d6-b75f-8bf6b84687f0
          name: Apache Log4j API
          identifiers:
            - idType: PURL
              idValue: pkg:maven/org.apache.logging.log4j/log4j-api
        - uuid: b844c9bd-55d6-478c-af59-954a932b6ad3
          name: Apache Log4j Core
          identifiers:
            - idType: CPE
              idValue: cpe:2.3:a:apache:log4j
            - idType: PURL
              idValue: pkg:maven/org.apache.logging.log4j/log4j-core

#
    # Reference to a component, in some cases directly to a specific release
    #
    # The release reference (release UUID) is only used in cases where a product
    # name includes a version and this version of the product always includes
    # the same releases of the component.
...
    release:
      type: object
      description: A TEA Component Release
      properties:
        uuid:
          description: A unique identifier for the TEA Component Release
          "$ref": "#/components/schemas/uuid"
        component:
          description: UUID of the TEA Component this release belongs to
          "$ref": "#/components/schemas/uuid"
        componentName:
          description: Name of the TEA Component this release belongs to
          type: string
          example: tomcat
        version:
          description: Version number
          type: string
          example: 1.2.3
        createdDate:
          description: Timestamp when this Release was created in TEA (for sorting purposes)
          "$ref": "#/components/schemas/date-time"
        releaseDate:
          description: Timestamp of the release
          "$ref": "#/components/schemas/date-time"
        preRelease:
          type: boolean
          description: |
            A flag indicating pre-release (or beta) status.
            May be disabled after the creation of the release object, but can't be enabled after creation of an object.
        identifiers:
          type: array
          description: List of identifiers for the component
          items:
            "$ref": "#/components/schemas/identifier"
        distributions:
          type: array
          description: List of different formats of this component release
          items:
            "$ref": "#/components/schemas/release-distribution"
      required:
        - uuid
        - version
        - createdDate
      examples:
        # Apache Tomcat 11.0.7
        - uuid: 605d0ecb-1057-40e4-9abf-c400b10f0345
          version: "11.0.7"
          createdDate: 2025-05-07T18:08:00Z
          releaseDate: 2025-05-12T18:08:00Z
          identifiers:
            - idType: PURL
              idValue: pkg:maven/org.apache.tomcat/tomcat@11.0.7
          distributions:
            - distributionId: 6a0d58e1-4896-4f1e-83f7-5feb6c032537
              description: Core binary distribution, zip archive
              identifiers:
                - idType: PURL
                  idValue: pkg:maven/org.apache.tomcat/tomcat@11.0.6?type=zip
              checksums:
                - algType: SHA_256
                  algValue: 9da736a1cdd27231e70187cbc67398d29ca0b714f885e7032da9f1fb247693c1
              url: https://repo.maven.apache.org/maven2/org/apache/tomcat/tomcat/11.0.7/tomcat-11.0.6.zip
              signatureUrl: https://repo.maven.apache.org/maven2/org/apache/tomcat/tomcat/11.0.7/tomcat-11.0.6.zip.asc
            - distributionId: dba01c13-dd96-4928-be72-9c87ffa8cab8
              description: Core binary distribution, tar.gz archive
              identifiers:
                - idType: PURL
                  idValue: pkg:maven/org.apache.tomcat/tomcat@11.0.6?type=tar.gz
              checksums:
                - algType: SHA_256
                  algValue: 2fcece641c62ba1f28e1d7b257493151fc44f161fb391015ee6a95fa71632fb9
              url: https://repo.maven.apache.org/maven2/org/apache/tomcat/tomcat/11.0.7/tomcat-11.0.6.tar.gz
              signatureUrl: https://repo.maven.apache.org/maven2/org/apache/tomcat/tomcat/11.0.7/tomcat-11.0.6.tar.gz.asc
            - distributionId: cfe068c2-fae7-43d0-97ec-5ea092454040
              description: Core binary distribution, Windows x64 zip archive
              identifiers:
                - idType: PURL
                  idValue: pkg:maven/org.apache.tomcat/tomcat@11.0.6?classifier=windows-x64&type=zip
              checksums:
                - algType: SHA_256
                  algValue: 62a5c358d87a8ef21d7ec1b3b63c9bbb577453dda9c00cbb522b16cee6c23fc4
              url: https://repo.maven.apache.org/maven2/org/apache/tomcat/tomcat/11.0.7/tomcat-11.0.6-windows-x64.zip
              signatureUrl: https://repo.maven.apache.org/maven2/org/apache/tomcat/tomcat/11.0.7/tomcat-11.0.6.zip.asc
            - distributionId: de45ffaf-e4b5-47b5-be28-444a76df098e
              description: Core binary distribution, Windows Service Installer (MSI)
              checksums:
                - algType: SHA_512
                  algValue: 1d3824e7643c8aba455ab0bd9e67b14a60f2aaa6aa7775116bce40eb0579e8ced162a4f828051d3b867e96ee2858ec5da0cc654e83a83ba30823cbea0df4ff96
              url: https://dlcdn.apache.org/tomcat/tomcat-11/v11.0.7/bin/apache-tomcat-11.0.7.exe
              signatureUrl: https://downloads.apache.org/tomcat/tomcat-11/v11.0.7/bin/apache-tomcat-11.0.7.exe.asc
        # A pre-release of Apache Tomcat
        - uuid: 95f481df-f760-47f4-b2f2-f8b76d858450
          version: "11.0.0-M26"
          createdDate: 2024-09-13T17:49:00Z
          preRelease: true
          identifiers:
            - idType: PURL
              idValue: pkg:maven/org.apache.tomcat/tomcat@11.0.0-M26
    release-distribution:
      type: object
      properties:
        distributionId:
          description: A unique identifier for the TEA Distribution object
          "$ref": "#/components/schemas/uuid"
        description:
          type: string
          description: Free-text description of the distribution.
        identifiers:
          type: array
          description: List of identifiers specific to this distribution.
          items:
            $ref: "#/components/schemas/identifier"
        url:
          type: string
          description: Direct download URL for the distribution.
          format: url
        signatureUrl:
          type: string
          description: Direct download URL for the distribution's external signature.
          format: url
        checksums:
          type: array
          description: List of checksums for the distribution.
          items:
            "$ref": "#/components/schemas/checksum"
      required:
        - distributionId
      examples:
        - distributionId: 6a0d58e1-4896-4f1e-83f7-5feb6c032537
          description: Core binary distribution, zip archive
          identifiers:
            - idType: PURL
              idValue: pkg:maven/org.apache.tomcat/tomcat@11.0.6?type=zip
          checksums:
            - algType: SHA_256
              algValue: 9da736a1cdd27231e70187cbc67398d29ca0b714f885e7032da9f1fb247693c1
          url: https://repo.maven.apache.org/maven2/org/apache/tomcat/tomcat/11.0.7/tomcat-11.0.6.zip
          signatureUrl: https://repo.maven.apache.org/maven2/org/apache/tomcat/tomcat/11.0.7/tomcat-11.0.6.zip.asc
        - distributionId: dba01c13-dd96-4928-be72-9c87ffa8cab8
          description: Core binary distribution, tar.gz archive
          identifiers:
            - idType: PURL
              idValue: pkg:maven/org.apache.tomcat/tomcat@11.0.6?type=tar.gz
          checksums:
            - algType: SHA_256
              algValue: 2fcece641c62ba1f28e1d7b257493151fc44f161fb391015ee6a95fa71632fb9
          url: https://repo.maven.apache.org/maven2/org/apache/tomcat/tomcat/11.0.7/tomcat-11.0.6.tar.gz
          signatureUrl: https://repo.maven.apache.org/maven2/org/apache/tomcat/tomcat/11.0.7/tomcat-11.0.6.tar.gz.asc
        - distributionId: cfe068c2-fae7-43d0-97ec-5ea092454040
          description: Core binary distribution, Windows x64 zip archive
          identifiers:
            - idType: PURL
              idValue: pkg:maven/org.apache.tomcat/tomcat@11.0.6?classifier=windows-x64&type=zip
          checksums:
            - algType: SHA_256
              algValue: 62a5c358d87a8ef21d7ec1b3b63c9bbb577453dda9c00cbb522b16cee6c23fc4
          url: https://repo.maven.apache.org/maven2/org/apache/tomcat/tomcat/11.0.7/tomcat-11.0.6-windows-x64.zip
          signatureUrl: https://repo.maven.apache.org/maven2/org/apache/tomcat/tomcat/11.0.7/tomcat-11.0.6.zip.asc
        - distributionId: de45ffaf-e4b5-47b5-be28-444a76df098e
          description: Core binary distribution, Windows Service Installer (MSI)
          checksums:
            - algType: SHA_512
              algValue: 1d3824e7643c8aba455ab0bd9e67b14a60f2aaa6aa7775116bce40eb0579e8ced162a4f828051d3b867e96ee2858ec5da0cc654e83a83ba30823cbea0df4ff96
          url: https://dlcdn.apache.org/tomcat/tomcat-11/v11.0.7/bin/apache-tomcat-11.0.7.exe
          signatureUrl: https://downloads.apache.org/tomcat/tomcat-11/v11.0.7/bin/apache-tomcat-11.0.7.exe.asc
...
    artifact:
      type: object
      description: A security-related document
      properties:
        uuid:
          description: The UUID of the TEA Artifact object. Together with *version* uniquely identifies the TEA Artifact.
          "$ref": "#/components/schemas/uuid"
        version:
          description: |
            An integer with default value 1.
            Together with *uuid* uniquely identifies the TEA Artifact.
            This field can be used to designate successive, immutable revisions of an artefact content (e.g. an updated VEX file).
          type: integer
          default: 1
        name:
          type: string
          description: Name of TEA Artifact
        type:
          description: Type of TEA Artifact
          "$ref": "#/components/schemas/artifact-type"
        createdDate:
          description: The date when the TEA Artifact revision was created.
          "$ref": "#/components/schemas/date-time"
        distributionIds:
          type: array
          description: |
            List of TEA Component Release distributions that this TEA Artifact applies to.
            If absent or empty, the TEA Artifact applies to all distributions.
          items:
            "$ref": "#/components/schemas/uuid"
            description: |
              The distribution IDs of the TEA component release distributions that this TEA Artifact applies to.
        formats:
          type: array
          description: |
            List of objects with the same content, but in different formats.
            The order of the list has no significance.
          items:
            "$ref": "#/components/schemas/artifact-format"
      required:
        - uuid
        - type
        - formats
      examples:
        - uuid: 1cb47b95-8bf8-3bad-a5a4-0d54d86e10ce
          name: Build SBOM
          type: BOM
          formats:
            - mediaType: application/vnd.cyclonedx+xml
              description: CycloneDX SBOM (XML)
              url: https://repo.maven.apache.org/maven2/org/apache/logging/log4j/log4j-core/2.24.3/log4j-core-2.24.3-cyclonedx.xml
              signatureUrl: https://repo.maven.apache.org/maven2/org/apache/logging/log4j/log4j-core/2.24.3/log4j-core-2.24.3-cyclonedx.xml.asc
              checksums:
                - algType: MD5
                  algValue: 2e1a525afc81b0a8ecff114b8b743de9
                - algType: SHA-1
                  algValue: 5a7d4caef63c5c5ccdf07c39337323529eb5a770
        - uuid: dfa35519-9734-4259-bba1-3e825cf4be06
          version: 7
          name: Vulnerability Disclosure Report
          type: VULNERABILITIES
          formats:
            - mediaType: application/vnd.cyclonedx+xml
              description: CycloneDX VDR (XML)
              url: https://logging.apache.org/cyclonedx/vdr.xml
              checksums:
                - algType: SHA-256
                  algValue: 75b81020b3917cb682b1a7605ade431e062f7a4c01a412f0b87543b6e995ad2a
    artifact-type:
      type: string
      description: Specifies the type of external reference.
      enum:
        - ATTESTATION
        - BOM
        - BUILD_META
        - CERTIFICATION
        - FORMULATION
        - LICENSE
        - RELEASE_NOTES
        - SECURITY_TXT
        - THREAT_MODEL
        - VULNERABILITIES
        - OTHER
    artifact-format:
      type: object
      description: A security-related document in a specific format
      properties:
        mediaType:
          type: string
          description: The Media Type of the document
        description:
          type: string
          description: A free text describing the TEA Artifact
        url:
          type: string
          description: Direct download URL for the TEA Artifact
          format: url
        signatureUrl:
          type: string
          description: Direct download URL for an external signature of the TEA Artifact
          format: url
        checksums:
          type: array
          description: List of checksums for the TEA Artifact
          items:
            "$ref": "#/components/schemas/checksum"
    checksum:
      type: object
      properties:
        algType:
          description: Checksum algorithm
          "$ref": "#/components/schemas/checksum-type"
        algValue:
          type: string
          description: Checksum value
      required:
        - algType
        - algValue
    checksum-type:
      type: string
      description: Checksum algorithm
      enum:
        - MD5
        - SHA-1
        - SHA-256
        - SHA-384
        - SHA-512
        - SHA3-256
        - SHA3-384
        - SHA3-512
        - BLAKE2b-256
        - BLAKE2b-384
        - BLAKE2b-512
        - BLAKE3
    unknown-error-type:
      type: string
      description: Classification of TEA error response
      enum:
        - OBJECT_UNKNOWN
        - OBJECT_NOT_SHAREABLE
    #
    # Types used in API responses
    #
    error-response:
      type: object
      description: Error response
      additionalProperties: false
      properties:
        error:
          $ref: "#/components/schemas/unknown-error-type"
      required:
        - error
...
  responses:
...
    400-invalid-request:
      description: Request was Invalid
      content:
        application/json: {}
...
    404-object-by-id-not-found:
      description: Object requested by identifier not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/error-response"
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    basicAuth:
      type: http
      scheme: basic
security:
  - bearerAuth: []
  - basicAuth: []
...